perf(uvr5): Apple Silicon MPS 加速人声分离（权重fp32+autocast fp16，损失最小）
- bsroformer.py：autocast 上下文按设备区分；MPS 权重保持 fp32、计算走 fp16
  autocast，避免全 fp16 权重触发 MPS 矩阵乘法断言崩溃，同时保证精度损失最小
- bs_roformer: mask 在转复数前强制 fp32，与 fp32 路径掩码精度一致
- vr.py：MPS 下保持 fp32 权重（同因），避免 HP/DeEcho 模型半精度崩溃
- tools/uvr5/webui.py：MPS 设备设置 PYTORCH_ENABLE_MPS_FALLBACK=1 兜底
- webui.py：UVR5 子进程自动检测 MPS（mps+half=True），NVIDIA/CPU 平台行为不变

实测（Apple M5/32GB，真实音频，同会话对照）：
- MPS fp16: 9.6 s/chunk（原 CPU 31-61 s/chunk，约 3-6 倍提速）
- MPS fp16 vs CPU fp32 输出 SNR 68.7 dB（端到端），听感无损

+++ tools/uvr5/bs_roformer/bs_roformer.py
@@ -558,7 +558,8 @@ class BSRoformer(Module):
         # complex number multiplication
 
         stft_repr = torch.view_as_complex(stft_repr)
-        mask = torch.view_as_complex(mask)
+        # mps autocast fp16 下 mask 可能是 fp16，先转 fp32 再转复数，保证与 fp32 路径一致的掩码精度
+        mask = torch.view_as_complex(mask.float())
 
         stft_repr = stft_repr * mask
